Promjena PricePerHour i Amount podataka iz type int u type decimal, dodavanje enum-a Currency u ParkingPayments entitet te kreiranje migracije i nadogradnje db na temelju tih promjena

+++ ParkingSystem/Controllers/ParkingEntriesController.cs
@@ -107,7 +107,7 @@ namespace ParkingSystem.Controllers
 
             var duration = ticketExpiration.Value - ticketTakeover.Value;
 
-            var totalHours = Math.Ceiling(duration.TotalHours);
+            var totalHours = (decimal)Math.Ceiling(duration.TotalHours);
 
             var totalPrice = (int)(totalHours * price);
